logic for client-server basic architecture finished - added suport for when user exits application

+++ server/src/main/java/ui/controller/ServerController.java
@@ -1,11 +1,14 @@
 package ui.controller;
 
 import dto.impl.UserDTO;
+import java.awt.event.WindowAdapter;
+import java.awt.event.WindowEvent;
 import lombok.extern.java.Log;
 import server.Server;
 import ui.table.model.UsersTableModel;
 import ui.view.ServerForm;
 
+import javax.swing.*;
 import java.io.IOException;
 
 @Log
@@ -34,6 +37,19 @@ public class ServerController {
     private void setListeners() {
         serverForm.getBtnStartServer().addActionListener(a -> startServer());
         serverForm.getBtnStopServer().addActionListener(a -> stopServer());
+        serverForm.addWindowListener(new WindowAdapter() {
+            @Override
+            public void windowClosing(WindowEvent e) {
+               if(!shouldServerBeTerminated())
+                   serverForm.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
+               else{
+                   serverForm.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+                   if(server != null)
+                       server.terminate();
+                   log.info("server STOPPED");
+               }
+            }
+        });
     }
 
     private void startServer() {
@@ -49,20 +65,21 @@ public class ServerController {
     }
 
     private void stopServer() {
-         if(!server.getClients().isEmpty()){
-             serverForm.warningDialog("Postoje aktivne sesije","OPREZ");
-             if(serverForm.confirmDialog("Da li ste sigurni da zelite da ugasite server?", "OPREZ")){
-                 server.terminate();
-                 serverForm.serverStopped();
-                 log.info("server STOPPED");
-             }
-             return;
-         }
+        if(!shouldServerBeTerminated())
+            return;
         server.terminate();
         serverForm.serverStopped();
         log.info("server STOPPED");
     }
 
+    private boolean shouldServerBeTerminated(){
+        if(server != null && !server.getClients().isEmpty()){
+             serverForm.warningDialog("Postoje aktivne sesije","OPREZ");
+             return serverForm.confirmDialog("Da li ste sigurni da zelite da ugasite server?", "OPREZ");
+             }
+        return true;
+    }
+    
     public void addUser(UserDTO user){
         ((UsersTableModel)serverForm.getTblActiveUsers().getModel()).addUser(user);
     }
